+++ app/src/main/java/com/chico/myhomebookkeeping/db/entity/CashAccount.kt
@@ -15,7 +15,9 @@ data class CashAccount(
     @ColumnInfo(name = "is_cash_account_default")
     val isCashAccountDefault: Boolean?,
     @ColumnInfo(name = "cash_account_name_ru")
-    val accountNameRu: String? = null
+    val accountNameRu: String? = null,
+    @ColumnInfo(name = "cash_account_name_pl")
+    val accountNamePl: String? = null
     ) {
     @PrimaryKey(autoGenerate = true)
     var cashAccountId: Int? = null