+++ app/src/main/java/com/chico/myhomebookkeeping/ui/firstLaunch/FirstLaunchCategoriesFragment.kt
@@ -75,9 +75,11 @@ class FirstLaunchCategoriesFragment : Fragment(R.layout.fragment_first_launch_ca
                     FirstLaunchCategoryGroupItem(
                         parentName = it.group.parentName,
                         parentNameRu = it.group.parentNameRu,
+                        parentNamePl = it.group.parentNamePl,
                         isIncome = it.group.isIncome,
                         subcategories = it.group.subcategories,
-                        subcategoriesRu = it.group.subcategoriesRu
+                        subcategoriesRu = it.group.subcategoriesRu,
+                        subcategoriesPl = it.group.subcategoriesPl
                     )
                 }
         }
@@ -87,10 +89,10 @@ class FirstLaunchCategoriesFragment : Fragment(R.layout.fragment_first_launch_ca
         ) : RecyclerView.ViewHolder(binding.root) {
             fun bind(item: SelectableDefaultCategoryGroup) {
                 with(binding) {
-                    defaultCategoryName.text = if (languageTag == Constants.APP_LANGUAGE_RUSSIAN) {
-                        item.group.parentNameRu
-                    } else {
-                        item.group.parentName
+                    defaultCategoryName.text = when (languageTag) {
+                        Constants.APP_LANGUAGE_RUSSIAN -> item.group.parentNameRu
+                        Constants.APP_LANGUAGE_POLISH -> item.group.parentNamePl
+                        else -> item.group.parentName
                     }
                     defaultCategorySubcategoryCount.text = itemView.context.getString(
                         R.string.first_launch_categories_subcategory_count,