+++ app/src/main/java/com/chico/myhomebookkeeping/ui/firstLaunch/FirstLaunchSetupItem.kt
@@ -8,7 +8,9 @@ data class FirstLaunchSetupItem(
 data class FirstLaunchCategoryGroupItem(
     val parentName: String,
     val parentNameRu: String?,
+    val parentNamePl: String?,
     val isIncome: Boolean,
     val subcategories: List<String>,
-    val subcategoriesRu: List<String>
+    val subcategoriesRu: List<String>,
+    val subcategoriesPl: List<String>
 )